register method added

+++ src/main/resources/static/js/usuarios.js
@@ -30,6 +30,11 @@ async function cargarUsuarios() {
 }
 
 async function eliminarUsuario(id) {
+
+          if(!confirm('Desea eliminar este usuario?')) {
+            return;
+          }
+
           const request = await fetch('api/usuarios/' + id , {
             method: 'DELETE',
             headers: {
@@ -37,5 +42,5 @@ async function eliminarUsuario(id) {
               'Content-Type': 'application/json'
             },
           });
-
+          location.reload()
 }
